+++ sandbox/plc4c/spi/src/main/include/plc4c/types_private.h
